fix : 유틸함수 파일 분리
