+++ src/utils/localStorage.ts
@@ -1,0 +1,3 @@
+export const getLocalStorage = (storageKey: string): string | null => {
+  return window.localStorage.getItem(storageKey);
+};
